Move settings page to pages directory
